backend/requirements.txt
fastapi==0.111.2
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.32
psycopg[binary]==3.1.32
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.0
alembic==1.11.1
paho-mqtt==1.6.4
python-dotenv==1.0.0
